+++ tencentcloud/clb/v20180317/models.go
@@ -1253,7 +1253,7 @@ type CreateLoadBalancerRequestParams struct {
 	// 仅适用于公网负载均衡。CMCC | CTCC | CUCC，分别对应 移动 | 电信 | 联通，如果不指定本参数，则默认使用BGP。可通过 DescribeSingleIsp 接口查询一个地域所支持的Isp。如果指定运营商，则网络计费式只能使用按带宽包计费(BANDWIDTH_PACKAGE)。
 	VipIsp *string `json:"VipIsp,omitempty" name:"VipIsp"`
 
-	// 购买负载均衡的同时，给负载均衡打上标签。
+	// 购买负载均衡的同时，给负载均衡打上标签，最大支持20个标签键值对。
 	Tags []*TagInfo `json:"Tags,omitempty" name:"Tags"`
 
 	// 指定VIP申请负载均衡。指定此参数后：
@@ -1338,7 +1338,7 @@ type CreateLoadBalancerRequest struct {
 	// 仅适用于公网负载均衡。CMCC | CTCC | CUCC，分别对应 移动 | 电信 | 联通，如果不指定本参数，则默认使用BGP。可通过 DescribeSingleIsp 接口查询一个地域所支持的Isp。如果指定运营商，则网络计费式只能使用按带宽包计费(BANDWIDTH_PACKAGE)。
 	VipIsp *string `json:"VipIsp,omitempty" name:"VipIsp"`
 
-	// 购买负载均衡的同时，给负载均衡打上标签。
+	// 购买负载均衡的同时，给负载均衡打上标签，最大支持20个标签键值对。
 	Tags []*TagInfo `json:"Tags,omitempty" name:"Tags"`
 
 	// 指定VIP申请负载均衡。指定此参数后：