+++ tencentcloud/tcss/v20201101/models.go
@@ -10312,6 +10312,15 @@ type DescribePurchaseStateInfoResponseParams struct {
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	GivenAuthorizedCnt *uint64 `json:"GivenAuthorizedCnt,omitempty" name:"GivenAuthorizedCnt"`
 
+	// 起始时间
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	BeginTime *string `json:"BeginTime,omitempty" name:"BeginTime"`
+
+	// 子状态(具体意义依据State字段而定)
+	// State为4时，有效值为: ISOLATE(隔离) DESTROED(已销毁)
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	SubState *string `json:"SubState,omitempty" name:"SubState"`
+
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
 }